requirements.txt
cycler==0.10.0
Cython==0.28.1
Django==2.0.3
djangorestframework==3.7.7
kiwisolver==1.0.1
matplotlib==2.2.2
nltk==3.2.5
numpy==1.14.2
pandas==0.22.0
pyparsing==2.2.0
pystan==2.17.1.0
python-dateutil==2.7.2
pytz==2018.3
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
xlrd==1.1.0
